+++ CHANGELOG.md
@@ -61,3 +61,5 @@ Log de alterações da aplicação
 ### Versão 1.6
 * **Docker:** Adicionando scripts para criação do ambiente Docker da API.
 * **Email:** Corrigindo a palavra Software no template dos emails.
+* **Projetos:** Refatoração do metodo editar.
+* **Integrantes:** Refatoração dos metodos ex-integrantes e atuais-integrantes para trazer o titulo do projeto.